fix: queues-bytes-and-d1-prd-type (#33)
# Changes

## Queues
* When producers sends `Uint8Array` it gets transformed into `ArrayBuffer` on the receiving consumer, this checks and transforms it back to `Uint8Array`, this allows `CopyBytesToGo` to be used - fixed https://github.com/Darckfast/workers-go/issues/45

## D1
* `SQLDurationMs` and `Duration` were both using int64, while this works locally, when deployed on Cloudflare, D1 returns a float, which was resulting in a conversion error being returned - fixes https://github.com/Darckfast/workers-go/issues/46

Reviewed-on: https://codeberg.org/darckfast/workers-go/pulls/33

+++ platform/cloudflare/d1/v2/types.go
@@ -11,16 +11,16 @@ type Meta struct {
 	ServedBy       string `json:"served_by"`
 	ServedByRegion string `json:"served_by_region"`
 	Timings        struct {
-		SQLDurationMs int64 `json:"sql_duration_ms"`
+		SQLDurationMs float64 `json:"sql_duration_ms"`
 	} `json:"timings"`
-	Duration        int64 `json:"duration"`
-	Changes         int64 `json:"changes"`
-	LastRowID       int64 `json:"last_row_id"`
-	SizeAfter       int64 `json:"size_after"`
-	RowsRead        int64 `json:"rows_read"`
-	RowsWritten     int64 `json:"rows_written"`
-	ServedByPrimary bool  `json:"served_by_primary"`
-	ChangedDB       bool  `json:"changed_db"`
+	Duration        float64 `json:"duration"`
+	Changes         int64   `json:"changes"`
+	LastRowID       int64   `json:"last_row_id"`
+	SizeAfter       int64   `json:"size_after"`
+	RowsRead        int64   `json:"rows_read"`
+	RowsWritten     int64   `json:"rows_written"`
+	ServedByPrimary bool    `json:"served_by_primary"`
+	ChangedDB       bool    `json:"changed_db"`
 }
 
 //easyjson:json
